fix(ParticipantOnboarding): Add issuer and id in participant credential
This commit contains below changes,

1. Add issuer and id while generate the legal participant
2. Refactor ParticipantService class and move some methods to Signer service
