+++ wizard-core/src/main/java/eu/gaiax/wizard/core/service/signer/SignerService.java
@@ -9,6 +9,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import eu.gaiax.wizard.api.client.SignerClient;
 import eu.gaiax.wizard.api.exception.EntityNotFoundException;
 import eu.gaiax.wizard.api.model.*;
+import eu.gaiax.wizard.api.model.setting.ContextConfig;
 import eu.gaiax.wizard.api.utils.CommonUtils;
 import eu.gaiax.wizard.api.utils.S3Utils;
 import eu.gaiax.wizard.api.utils.Validate;
@@ -19,16 +20,20 @@ import eu.gaiax.wizard.dao.entity.participant.Participant;
 import eu.gaiax.wizard.dao.repository.participant.ParticipantRepository;
 import eu.gaiax.wizard.vault.Vault;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import java.io.File;
 import java.nio.charset.Charset;
-import java.util.Map;
-import java.util.UUID;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
+import java.time.format.DateTimeFormatter;
+import java.util.*;
 
 /**
  * The type Signer service.
@@ -38,11 +43,16 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class SignerService {
 
+    @Value("${wizard.host.wizard}")
+    private String wizardHost;
+    @Value("${wizard.gaiax.tnc}")
+    private String tnc;
+    private final ContextConfig contextConfig;
     private final CredentialService credentialService;
     private final ParticipantRepository participantRepository;
     private final SignerClient signerClient;
     private final S3Utils s3Utils;
-    private final ObjectMapper objectMapper;
+    private final ObjectMapper mapper;
     private final ScheduleService scheduleService;
     private final Vault vault;
 
@@ -50,7 +60,63 @@ public class SignerService {
         log.info("SignerService(createParticipantJson) -> Initiate the legal participate creation process for participant {}", participantId);
         Participant participant = this.participantRepository.findById(participantId).orElse(null);
         Validate.isNull(participant).launch(new EntityNotFoundException("participant.not.found"));
+        this.prepareCredentialSubjectForLegalParticipant(participant);
         this.createParticipantJson(participant, participant.getId().toString(), participant.isOwnDidSolution());
+    }
+
+    @SneakyThrows
+    private Map<String, Object> prepareCredentialSubjectForLegalParticipant(Participant participant) {
+        log.info("ParticipantService(prepareCredentialSubjectForLegalParticipant) -> Prepare credential subject for signer tool.");
+        TypeReference<Map<String, Object>> typeReference = new TypeReference<>() {
+        };
+        Map<String, Object> credential = this.mapper.readValue(participant.getCredential(), typeReference);
+        Map<String, Object> legalParticipant = this.mapper.convertValue(credential.get("legalParticipant"), typeReference);
+        Map<String, Object> legalRegistrationNumber = this.mapper.convertValue(credential.get("legalRegistrationNumber"), typeReference);
+        //Add @context in the credential
+        legalParticipant.put("@context", this.contextConfig.participant());
+        legalParticipant.put("type", List.of("VerifiableCredential"));
+        legalParticipant.put("id", participant.getDid());
+        legalParticipant.put("issuer", participant.getDid());
+        String issuanceDate = LocalDateTime.now().atZone(ZoneOffset.UTC).format(DateTimeFormatter.ISO_OFFSET_DATE_TIME);
+        legalParticipant.put("issuanceDate", issuanceDate);
+
+        Map<String, Object> participantCredentialSubject = this.mapper.convertValue(legalParticipant.get("credentialSubject"), typeReference);
+        participantCredentialSubject.put("id", this.formParticipantJsonUrl(participant.getId()) + "#0");
+        participantCredentialSubject.put("type", "gx:LegalParticipant");
+        String registrationId = this.formParticipantJsonUrl(participant.getId()) + "#1";
+        participantCredentialSubject.put("gx:legalRegistrationNumber", Map.of("id", registrationId));
+
+        legalParticipant.put("credentialSubject", participantCredentialSubject);
+
+        legalRegistrationNumber.put("@context", this.contextConfig.registrationNumber());
+        legalRegistrationNumber.put("type", List.of("gx:legalRegistrationNumber"));
+        legalRegistrationNumber.put("id", registrationId);
+
+        Map<String, Object> tncVc = new TreeMap<>();
+        tncVc.put("@context", this.contextConfig.tnc());
+        tncVc.put("type", List.of("VerifiableCredential"));
+        tncVc.put("id", participant.getDid());
+        tncVc.put("issuer", participant.getDid());
+        tncVc.put("issuanceDate", issuanceDate);
+
+        Map<String, Object> tncCredentialSubject = new HashMap<>();
+        tncCredentialSubject.put("type", "gx:GaiaXTermsAndConditions");
+        tncCredentialSubject.put("@Context", this.contextConfig.tnc());
+        tncCredentialSubject.put("id", this.formParticipantJsonUrl(participant.getId()) + "#2");
+        tncCredentialSubject.put("gx:termsAndConditions", this.tnc.replaceAll("\\\\n", "\n"));
+
+        tncVc.put("credentialSubject", tncCredentialSubject);
+
+        credential.put("legalParticipant", legalParticipant);
+        credential.put("legalRegistrationNumber", legalRegistrationNumber);
+        credential.put("gaiaXTermsAndConditions", tncVc);
+        participant.setCredential(this.mapper.writeValueAsString(credential));
+        log.info("ParticipantService(prepareCredentialSubjectForLegalParticipant) -> CredentialSubject has been created successfully.");
+        return credential;
+    }
+
+    private String formParticipantJsonUrl(UUID participantId) {
+        return this.wizardHost + participantId.toString() + "/participant.json";
     }
 
     public void createParticipantJson(Participant participant, String key, boolean ownDid) {
@@ -65,16 +131,16 @@ public class SignerService {
             }
             TypeReference<Map<String, Object>> typeReference = new TypeReference<>() {
             };
-            Map<String, Object> credentials = this.objectMapper.readValue(participant.getCredential(), typeReference);
+            Map<String, Object> credentials = this.mapper.readValue(participant.getCredential(), typeReference);
             CreateVCRequest request = new CreateVCRequest(HashingService.encodeToBase64(privateKey), participant.getDid(), participant.getDid(), credentials);
             log.info("SignerService(createParticipantJson) -> Initiate the signer client call to create legal participant json.");
             ResponseEntity<Map<String, Object>> responseEntity = this.signerClient.createVc(request);
             log.info("SignerService(createParticipantJson) -> Receive success response from signer tool.");
-            String participantString = this.objectMapper.writeValueAsString(responseEntity.getBody().get("data"));
+            String participantString = this.mapper.writeValueAsString(responseEntity.getBody().get("data"));
             FileUtils.writeStringToFile(file, participantString, Charset.defaultCharset());
             String hostedPath = participant.getId() + "/participant.json";
             this.s3Utils.uploadFile(hostedPath, file);
-            this.credentialService.createCredential(participantString, hostedPath, CredentialTypeEnum.LEGAL_PARTICIPANT.getCredentialType(), null, participant);
+            this.credentialService.createCredential(participantString, this.formParticipantJsonUrl(participant.getId()), CredentialTypeEnum.LEGAL_PARTICIPANT.getCredentialType(), null, participant);
             participant.setStatus(RegistrationStatus.PARTICIPANT_JSON_CREATED.getStatus());
             log.debug("SignerService(createParticipantJson) -> Participant json has been created for participant {} with selfDescription {}", participant.getId(), participantString);
         } catch (Exception e) {
@@ -99,7 +165,7 @@ public class SignerService {
             log.info("SignerService(createDid) -> Initiated signerClient call for create did for domain {}", domain);
             ResponseEntity<Map<String, Object>> responseEntity = this.signerClient.createDid(createDidRequest);
             log.info("SignerService(createDid): -> Response has been received from signerClient for domain {}", domain);
-            String didString = this.objectMapper.writeValueAsString(((Map<String, Object>) responseEntity.getBody().get("data")).get("did"));
+            String didString = this.mapper.writeValueAsString(((Map<String, Object>) responseEntity.getBody().get("data")).get("did"));
             FileUtils.writeStringToFile(file, didString, Charset.defaultCharset());
             this.s3Utils.uploadFile(participant.getId() + "/did.json", file);
             participant.setStatus(RegistrationStatus.DID_JSON_CREATED.getStatus());
